+++ deno_dist/middleware.ts
@@ -2,7 +2,6 @@
 export * from './middleware/basic-auth/index.ts'
 export * from './middleware/bearer-auth/index.ts'
 export * from './middleware/body-parse/index.ts'
-export * from './middleware/cookie/index.ts'
 export * from './middleware/cors/index.ts'
 export * from './middleware/etag/index.ts'
 export * from './middleware/html/index.ts'
